requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via microsoft-kiota-authentication-azure
aiosignal==1.3.2
    # via aiohttp
anyio==4.9.0
    # via httpx
async-timeout==5.0.1
    # via aiohttp
attrs==25.3.0
    # via aiohttp
azure-core==1.32.0
    # via
    #   azure-identity
    #   microsoft-kiota-authentication-azure
azure-identity==1.21.0
    # via
    #   graphreporter (pyproject.toml)
    #   msgraph-sdk
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
commonmark==0.9.1
    # via rich
cryptography==44.0.2
    # via
    #   azure-identity
    #   msal
    #   pyjwt
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
et-xmlfile==2.0.0
    # via openpyxl
exceptiongroup==1.2.2
    # via anyio
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   microsoft-kiota-http
    #   msgraph-core
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.6.1
    # via opentelemetry-api
microsoft-kiota-abstractions==1.9.2
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
    #   msgraph-core
microsoft-kiota-authentication-azure==1.9.2
    # via msgraph-core
microsoft-kiota-http==1.9.2
    # via msgraph-core
microsoft-kiota-serialization-form==1.9.2
    # via msgraph-sdk
microsoft-kiota-serialization-json==1.9.2
    # via msgraph-sdk
microsoft-kiota-serialization-multipart==1.9.2
    # via msgraph-sdk
microsoft-kiota-serialization-text==1.9.2
    # via msgraph-sdk
msal==1.32.0
    # via
    #   graphreporter (pyproject.toml)
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
msgraph-core==1.3.2
    # via msgraph-sdk
msgraph-sdk==1.25.0
    # via graphreporter (pyproject.toml)
multidict==6.2.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.4
    # via pandas
openpyxl==3.0.10
    # via graphreporter (pyproject.toml)
opentelemetry-api==1.31.1
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.31.1
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.52b1
    # via opentelemetry-sdk
pandas==1.5.3
    # via graphreporter (pyproject.toml)
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pycparser==2.22
    # via cffi
pydantic==1.10.21
    # via graphreporter (pyproject.toml)
pygments==2.19.1
    # via rich
pyjwt==2.10.1
    # via msal
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==0.21.1
    # via graphreporter (pyproject.toml)
pytz==2025.1
    # via pandas
requests==2.32.3
    # via
    #   graphreporter (pyproject.toml)
    #   azure-core
    #   msal
rich==12.6.0
    # via graphreporter (pyproject.toml)
six==1.17.0
    # via
    #   azure-core
    #   python-dateutil
sniffio==1.3.1
    # via anyio
std-uritemplate==2.0.3
    # via microsoft-kiota-abstractions
typer==0.7.0
    # via graphreporter (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   azure-identity
    #   multidict
    #   opentelemetry-sdk
    #   pydantic
urllib3==2.3.0
    # via requests
wrapt==1.17.2
    # via deprecated
yarl==1.18.3
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
